Commit_AllMaster_17-10-22
