+++ Infrastructure/KSAA.Master/KSAA.Master.Infrastructure.Shared/Services/TaxCodeServices.cs
@@ -1,5 +1,5 @@
 ﻿using AutoMapper;
-using KSAA.Domain.Entities;
+using KSAA.Domain.Entities.Master;
 using KSAA.Domain.Interfaces.Repositories;
 using KSAA.Master.Application.DTOs.Master;
 using KSAA.Master.Application.Features.Commands.TaxCodeCommand;
