+++ Shared/KSAA.Domain/Entities/Master/Company.cs
@@ -5,17 +5,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace KSAA.Domain.Entities
+namespace KSAA.Domain.Entities.Master
 {
-    public class Location : BaseEntity
+    public class Company : BaseEntity
     {
-        public virtual string? Location_Code { get; set; }
-        public virtual string? Address { get; set; }
-        public virtual string? GSTRegistrationNo { get; set; }
-        public virtual string? TypeOfUnit { get; set; }
-        public virtual string? ProductsManufactured { get; set; }
-        public virtual string? ProductsTraded { get; set; }
-        public virtual string? TypeOfServicesProvided { get; set; }
+        public string? Company_Name { get; set; }
+        public string? Company_Code { get; set; }
+        public string? Company_Address { get; set; }
         public virtual string? IP { get; set; }
         public virtual string? BrowserCase { get; set; }
         public virtual string? CreatedBy { get; set; }
