+++ Shared/KSAA.Domain/Entities/Master/CustomerCode.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace KSAA.Domain.Entities
+namespace KSAA.Domain.Entities.Master
 {
     public class CustomerCode : BaseEntity
     {
