+++ Shared/KSAA.Domain/Entities/Master/GLIncome_Mapping.cs
@@ -5,15 +5,13 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace KSAA.Domain.Entities
+namespace KSAA.Domain.Entities.Master
 {
-    public class TBTagging : BaseEntity
+    public class GLIncome_Mapping : BaseEntity
     {
-        public virtual string? TBTaggingCode { get; set; }
-        public virtual string? GLCode { get; set; }
-        public virtual string? GLName { get; set; }
-        public virtual decimal Amount { get; set; }
-        public virtual string? TagCode { get; set; }
+        public virtual string? GLIncomeCode { get; set; }
+        public virtual string? GLIncomeDescription { get; set; }
+        public virtual string? CreatedBy { get; set; }
         public virtual DateTime? CreatedOn { get; set; }
         public virtual string? ModifiedBy { get; set; }
         public virtual DateTime? ModifiedOn { get; set; }
