+++ Shared/KSAA.Domain/Entities/Master/PlantCode.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace KSAA.Domain.Entities
+namespace KSAA.Domain.Entities.Master
 {
     public class PlantCode : BaseEntity
     {
