+++ Shared/KSAA.Domain/Entities/Master/VendorCode.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace KSAA.Domain.Entities
+namespace KSAA.Domain.Entities.Master
 {
     public class VendorCode : BaseEntity
     {
